requirements.txt
flask==2.3.3
python-telegram-bot==20.4
python-dotenv==1.0.0
requests==2.31.0
pymongo==4.5.0
Pillow==10.0.0